requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.2
requests==2.32.3
user-agents==2.2.0
python-dotenv==1.0.1
